connectjcuServer/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
idna==3.6
mysqlclient==2.2.4
packaging==23.2
pillow==10.2.0
PyJWT==2.8.0
PyMySQL==1.1.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.2.1
